fix: options IA_NA and IA_PD are not received with solicit message
Initially when dibbler client is started, both options IA_NA and
IA_PD timestamps are initialized at the same time. So the initial
solicit message does have both options.

When Reply msg is received from the server (i.e. after exchanging
Advertise and Request msgs), the IN_NA timestamp is reinitialized,
but the IA_PD timestamp is not initialized. Since the reply msg is
normally received after 1-2 secs, the IN_NA timestamp differs from
the IA_PD timestamp by 1-2 secs.

Though the timeout is the same for options, due to the changes in the
timestamp, the next msg RENEW is sent with IA_PD first and then after
~2 secs another RENEW msg is sent with option IA_NA.

This causes the next solicit message (test case: DHCPv6 server is
disconnected and RENEW fails multiple times) to be sent separately
for IA_NA and IA_PD options.

Fix by initializing IA_PD timestamp in addPrefix(), which is called
when the Reply message is received.

Ref: cdrouter dhcpv6_pd_100 ( OWF-1487 )